requirements.txt
aws-cdk.core
aws-cdk.aws_ec2
aws-cdk.aws_ecs
aws-cdk.aws_ecr
aws-cdk.aws_elasticloadbalancingv2
aws-cdk.aws-dynamodb
